Move business code to package
